PostBlog/requirements.txt
amqp==5.1.1
asgiref==3.5.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.12.0
django-filter==21.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.3.7
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.1.0
prompt-toolkit==3.0.29
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
redis==2.10.3
requests==2.27.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.0
